requirements.txt
google-auth==2.36.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
google-auth-httplib2

